[srp-server] replace SRP Host pointer with use SRP service update ID (#6407)
We are using the SRP Host pointer for indexing an outstanding SRP
update when the Advertising Proxy result is passed back with
otSrpServerHandleServiceUpdateResult. There are chances that an old,
freed Host object may be matched again with a newly allocated Host
object which coincidentally has the same address with the old Host
object.

This commit solves this issue by using a random generated uint32_t SRP
service update ID for indexing outstanding updates.