[posix] recreate connection after hard reset RCP(USB CDC ACM) device
When RCP(USB CDC ACM) device performs hard reset the communication
is lost because the connection is broken.

This commit handles this situation by recreation the connection after
resetting the RCP device.

It uses udev library to check if device enumeration has ended and is
already added to host OS.

It works based on symlink to the RCP device created by udev
in /dev/serial/by-id/ directory.

This functionality is enabled with the option:
OPENTHREAD_SPINEL_CONFIG_RESET_CONNECTION=1