Fix for OPENTHREAD-1155.
In this commit, I have updated the OpenThread stack to support a Max Channel Page value of 255 when working with proprietary radio configurations.

Secondly, I have also modified the (<=) usage in two ‘if’ conditions as these operations as written today (with constants), are resulting in the compile time warnings (and cmake errors) specifically when working with proprietary radio configurations and setting Channel Min to 0.